requirements.txt
asgiref==3.4.1
Django==3.2.5
django-ckeditor==6.1.0
django-heroku==0.3.1
django-js-asset==1.2.2
gunicorn==20.1.0
psycopg2==2.9.1
python-decouple==3.4
pytz==2021.1
sqlparse==0.4.1
typing-extensions==3.10.0.0
whitenoise==5.3.0